+++ sdk/resourcemanager/securityinsights/armsecurityinsights/zz_generated_entityrelations_client.go
@@ -56,7 +56,7 @@ func NewEntityRelationsClient(subscriptionID string, credential azcore.TokenCred
 
 // GetRelation - Gets an entity relation.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-05-01-preview
+// Generated from API version 2022-07-01-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // workspaceName - The name of the workspace.
 // entityID - entity ID
@@ -106,7 +106,7 @@ func (client *EntityRelationsClient) getRelationCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-05-01-preview")
+	reqQP.Set("api-version", "2022-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil