+++ sdk/resourcemanager/securityinsights/armsecurityinsights/zz_generated_sourcecontrol_client.go
@@ -56,7 +56,7 @@ func NewSourceControlClient(subscriptionID string, credential azcore.TokenCreden
 
 // NewListRepositoriesPager - Gets a list of repositories metadata.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-05-01-preview
+// Generated from API version 2022-07-01-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // workspaceName - The name of the workspace.
 // repoType - The repo type.
@@ -110,7 +110,7 @@ func (client *SourceControlClient) listRepositoriesCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-05-01-preview")
+	reqQP.Set("api-version", "2022-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, repoType)